Refactor popover tests
