+++ Example/PopoverExampleTests/PopoverExampleTests.swift
@@ -1,0 +1,35 @@
+//
+//  PopoverExampleTests.swift
+//  PopoverExampleTests
+//
+//  Created by corin8823 on 1/8/17.
+//  Copyright © 2017 corin8823. All rights reserved.
+//
+
+import XCTest
+
+class PopoverExampleTests: XCTestCase {
+    
+    override func setUp() {
+        super.setUp()
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+    }
+    
+    override func tearDown() {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+        super.tearDown()
+    }
+    
+    func testExample() {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    }
+    
+    func testPerformanceExample() {
+        // This is an example of a performance test case.
+        self.measure {
+            // Put the code you want to measure the time of here.
+        }
+    }
+    
+}
